refactor: index foods feature specific components
