+++ src/features/foods/components/index.ts
@@ -1,2 +1,9 @@
+export { AddButton } from './AddButton';
+export { AddDrawer } from './AddDrawer';
+export { BadgeList } from './BadgeList';
 export { CategoryBadge } from './CategoryBadge';
 export { FilterBadge } from './FilterBadge';
+export { FilterButton } from './FilterButton';
+export { FoodList } from './FoodList';
+export { FoodSort } from './FoodSort';
+export { SearchBar } from './SearchBar';
